requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
docutils==0.19
gunicorn==20.1.0
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
pipreqs==0.4.13
Pygments==2.15.1
PyJWT==1.7.1
pymongo==3.12.3
PyMySQL==1.0.3
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
yarg==0.1.9
